+++ arch/mips64/mm/r4xx0.c
@@ -1750,7 +1750,7 @@ out:
  */
 static void r4k_flush_page_to_ram_s16d16i16(struct page * page)
 {
-	unsigned long addr = page_address(page) & PAGE_MASK;
+	unsigned long addr = (unsigned long) page_address(page) & PAGE_MASK;
 
 	if ((addr >= KSEG0 && addr < KSEG1) || (addr >= KSEG2)) {
 #ifdef DEBUG_CACHE
@@ -1762,7 +1762,7 @@ static void r4k_flush_page_to_ram_s16d16i16(struct page * page)
 
 static void r4k_flush_page_to_ram_s32d16i16(struct page * page)
 {
-	unsigned long addr = page_address(page) & PAGE_MASK;
+	unsigned long addr = (unsigned long) page_address(page) & PAGE_MASK;
 
 	if ((addr >= KSEG0 && addr < KSEG1) || (addr >= KSEG2)) {
 #ifdef DEBUG_CACHE
@@ -1774,7 +1774,7 @@ static void r4k_flush_page_to_ram_s32d16i16(struct page * page)
 
 static void r4k_flush_page_to_ram_s64d16i16(struct page * page)
 {
-	unsigned long addr = page_address(page) & PAGE_MASK;
+	unsigned long addr = (unsigned long) page_address(page) & PAGE_MASK;
 
 	if ((addr >= KSEG0 && addr < KSEG1) || (addr >= KSEG2)) {
 #ifdef DEBUG_CACHE
@@ -1786,7 +1786,7 @@ static void r4k_flush_page_to_ram_s64d16i16(struct page * page)
 
 static void r4k_flush_page_to_ram_s128d16i16(struct page * page)
 {
-	unsigned long addr = page_address(page) & PAGE_MASK;
+	unsigned long addr = (unsigned long) page_address(page) & PAGE_MASK;
 
 	if ((addr >= KSEG0 && addr < KSEG1) || (addr >= KSEG2)) {
 #ifdef DEBUG_CACHE
@@ -1798,7 +1798,7 @@ static void r4k_flush_page_to_ram_s128d16i16(struct page * page)
 
 static void r4k_flush_page_to_ram_s32d32i32(struct page * page)
 {
-	unsigned long addr = page_address(page) & PAGE_MASK;
+	unsigned long addr = (unsigned long) page_address(page) & PAGE_MASK;
 
 	if ((addr >= KSEG0 && addr < KSEG1) || (addr >= KSEG2)) {
 #ifdef DEBUG_CACHE
@@ -1810,7 +1810,7 @@ static void r4k_flush_page_to_ram_s32d32i32(struct page * page)
 
 static void r4k_flush_page_to_ram_s64d32i32(struct page * page)
 {
-	unsigned long addr = page_address(page) & PAGE_MASK;
+	unsigned long addr = (unsigned long) page_address(page) & PAGE_MASK;
 
 	if ((addr >= KSEG0 && addr < KSEG1) || (addr >= KSEG2)) {
 #ifdef DEBUG_CACHE
@@ -1822,7 +1822,7 @@ static void r4k_flush_page_to_ram_s64d32i32(struct page * page)
 
 static void r4k_flush_page_to_ram_s128d32i32(struct page * page)
 {
-	unsigned long addr = page_address(page) & PAGE_MASK;
+	unsigned long addr = (unsigned long) page_address(page) & PAGE_MASK;
 
 	if ((addr >= KSEG0 && addr < KSEG1) || (addr >= KSEG2)) {
 #ifdef DEBUG_CACHE
@@ -1834,7 +1834,7 @@ static void r4k_flush_page_to_ram_s128d32i32(struct page * page)
 
 static void r4k_flush_page_to_ram_d16i16(struct page * page)
 {
-	unsigned long addr = page_address(page) & PAGE_MASK;
+	unsigned long addr = (unsigned long) page_address(page) & PAGE_MASK;
 
 	if ((addr >= KSEG0 && addr < KSEG1) || (addr >= KSEG2)) {
 		unsigned long flags;
@@ -1850,7 +1850,7 @@ static void r4k_flush_page_to_ram_d16i16(struct page * page)
 
 static void r4k_flush_page_to_ram_d32i32(struct page * page)
 {
-	unsigned long addr = page_address(page) & PAGE_MASK;
+	unsigned long addr = (unsigned long) page_address(page) & PAGE_MASK;
 
 	if ((addr >= KSEG0 && addr < KSEG1) || (addr >= KSEG2)) {
 		unsigned long flags;
